[Refactor] change error handling, params and return type

+++ backend/Pocketfolio/src/main/java/com/ssafy/pocketfolio/api/service/PortfolioServiceImpl.java
@@ -85,14 +85,19 @@ public class PortfolioServiceImpl implements PortfolioService{
         log.debug("[GET] Service - findPortfolioList");
         List<PortfolioRes> portfolioRes = new ArrayList<>();
         
-        User user = userRepository.findById(userSeq).orElseThrow(() -> new IllegalArgumentException("해당 사용자가 존재하지 않습니다."));
-        List<Portfolio> portfolios = portfolioRepository.findAllByUser(user);
+        try {
+            User user = userRepository.findById(userSeq).orElseThrow(() -> new IllegalArgumentException("해당 사용자가 존재하지 않습니다."));
+            List<Portfolio> portfolios = portfolioRepository.findAllByUser(user);
 
-        for (Portfolio portfolio : portfolios) {
-            List<PortfolioUrl> urls = portfolioUrlRepository.findAllByPortfolio(portfolio);
-            List<Tag> tags = tagRepository.findAllByPortfolio(portfolio);
-            PortfolioRes result = PortfolioRes.toDto(portfolio, urls, tags);
-            portfolioRes.add(result);
+            for (Portfolio portfolio : portfolios) {
+                List<PortfolioUrl> urls = portfolioUrlRepository.findAllByPortfolio(portfolio);
+                List<Tag> tags = tagRepository.findAllByPortfolio(portfolio);
+                PortfolioRes result = PortfolioRes.toDto(portfolio, urls, tags);
+                portfolioRes.add(result);
+            }
+        } catch (Exception e) {
+            log.error(e.getMessage());
+            portfolioRes = null;
         }
 
         return portfolioRes;
@@ -117,10 +122,12 @@ public class PortfolioServiceImpl implements PortfolioService{
         log.debug("[PATCH] Service - updatePortfolio");
 
         Portfolio portfolio = portfolioRepository.findById(portSeq).orElseThrow(() -> new IllegalArgumentException("해당 포트폴리오가 존재하지 않습니다."));
-        if (portfolio.getUser().getUserSeq() != userSeq) {
+        
+        if (userSeq != portfolio.getUser().getUserSeq()) {
             log.error("권한 없음");
             return null;
         }
+        
         log.debug("portfolio" + portfolio);
         // 저장된 썸네일 주소
         String thumbnailUrl = portfolio.getThumbnail();
@@ -133,7 +140,7 @@ public class PortfolioServiceImpl implements PortfolioService{
             }
             thumbnailUrl = fileHandler.saveThumbnail(thumbnail, "portfolio" + File.separator + "thumbnail");
             if(thumbnailUrl == null) {
-                log.error("권한 없음");
+                log.error("썸네일 저장 실패");
                 return null;
             }
         } else {
@@ -165,10 +172,17 @@ public class PortfolioServiceImpl implements PortfolioService{
 
     // 포트폴리오 삭제
     @Override
+    @Transactional
     public Boolean deletePortfolio(long userSeq, long portSeq) {
         log.debug("[DELETE] Service - deletePortfolio");
         try {
             Portfolio portfolio = portfolioRepository.findById(portSeq).orElseThrow(() -> new IllegalArgumentException("해당 포트폴리오가 존재하지 않습니다."));
+
+            if (userSeq != portfolio.getUser().getUserSeq()) {
+                log.error("권한 없음");
+                return null;
+            }
+
             // 썸네일 삭제
             if (portfolio.getThumbnail() != null) {
                 fileHandler.deleteFile(portfolio.getThumbnail());
@@ -209,7 +223,7 @@ public class PortfolioServiceImpl implements PortfolioService{
         }
     }
 
-
+    // 파일 저장
     public void saveUrls(List<MultipartFile> files, Portfolio portfolio) throws IOException {
         for (MultipartFile file : files) {
             File dest = fileHandler.saveFile(file, "portfolio");
@@ -220,5 +234,4 @@ public class PortfolioServiceImpl implements PortfolioService{
             portfolioUrlRepository.save(url);
         }
     }
-
 }
